# Add a few microoptimizations to the generation code